Add Minimax chat completion. Add AnthropicFriendlyProvider.
